backend/requirements.txt
fastapi
uvicorn
python-jose
python-multipart
pyodbc
ldap3
python-dotenv
pydantic
pydantic-settings